fix: preserve serial test isolation
